Relocated is_angle_linear() from zmat to common
